@@ -7,7 +7,7 @@
     ],
     "license": "MIT",
     "require": {
-        "laravel/framework": "5.1.*",
+        "laravel/framework": "5.2.*",
         "caseyamcl/Phpoaipmh": "dev-master",
         "intervention/image": "dev-master",
         "guzzlehttp/guzzle": "dev-master",
@@ -15,7 +15,7 @@
         "rtconner/laravel-tagging": "~2.1",
         "maatwebsite/excel": "~2.1.0",
         "rap2hpoutre/laravel-log-viewer": "dev-master",
-        "zizaco/entrust": "1.4.1",
+        "zizaco/entrust": "5.2.x-dev",
         "roumen/sitemap": "dev-master",
         "cviebrock/image-validator": "dev-master",
         "anahkiasen/former": "^4.0",
@@ -52,7 +52,9 @@
         "preferred-install": "dist"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.0"
+        "phpunit/phpunit": "~4.0",
+        "symfony/dom-crawler": "~3.0",
+        "symfony/css-selector": "~3.0"
     },
     "autoload-dev": {
         "classmap": [
